celery-reflex-tutorial/requirements.txt
supabase
celery
redis
openai>=1.0.0
python-dotenv
reflex
requests
reflex
reflex==0.8.2